RISC-V: KVM: Allow zicfiss/zicfilp exts for Guest/VM
Extend the KVM ISA extension ONE_REG interface to allow KVM user space
to detect and enable zicfiss/zicfilp exts for Guest/VM, the rules defined
in the spec are as follows:
---
1) Zicfiss extension introduces the SSE field (bit 3) in henvcfg. If the SSE
field is set to 1, the Zicfiss extension is activated in VS-mode. When the SSE
field is 0, the Zicfiss extension remains inactive in VS-mode.

2) Zicfilp extension introduces the LPE field (bit 2) in henvcfg. When the LPE
field is set to 1, the Zicfilp extension is enabled in VS-mode. When the LPE
field is 0, the Zicfilp extension is not enabled in VS-mode.

Signed-off-by: Quan Zhou <zhouquan@iscas.ac.cn>